+++ ProjectManager/ProjectManager/Model/CardModel/CardEntity.swift
